Concatenate monasca-ui settings into horizon's settings

+++ cookbooks/devstack/recipes/mon-ui.rb
@@ -12,6 +12,9 @@ end
 # Use 'execute' resource because chef does not support symlinking directories
 execute "ln -sfv /usr/local/lib/python2.7/dist-packages/enabled/* /opt/stack/horizon/openstack_dashboard/local/enabled/"
 execute "ln -sv /usr/local/lib/python2.7/dist-packages/monitoring /opt/stack/horizon/monitoring"
+if ::File.exists?("/usr/local/lib/python2.7/dist-packages/monitoring/local_settings.py")
+    execute "cat /usr/local/lib/python2.7/dist-packages/monitoring/local_settings.py >> /opt/stack/horizon/openstack_dashboard/local/local_settings.py"
+end
 
 # install grafana and integrate with horizon
 if !::File.exists?("/usr/local/lib/python2.7/dist-packages/monitoring/static/grafana")
